+++ app/controllers/api/tasks_controller.rb
@@ -1,5 +1,9 @@
 class Api::TasksController < ApiController
 
+  def show
+    @task = current_task
+  end
+
   def update
     @task = current_task
     @task.assign_attributes update_task_params